QREPO-193 refactor based on review, add licence, use property for version, rename classes, use xml file as expected result
