+++ dspace-api/pom.xml
@@ -19,6 +19,7 @@
     <properties>
         <!-- This is the path to the root [dspace-src] directory. -->
         <root.basedir>${basedir}/..</root.basedir>
+        <dataformat.version>2.12.7</dataformat.version>
     </properties>
 
     <!--
@@ -626,7 +627,7 @@
         <dependency>
             <groupId>com.fasterxml.jackson.dataformat</groupId>
             <artifactId>jackson-dataformat-xml</artifactId>
-            <version>2.12.7</version>
+            <version>${dataformat.version}</version>
             <exclusions>
                 <exclusion>
                     <groupId>com.fasterxml.woodstox</groupId>
