+++ dspace-api/src/main/java/org/dspace/app/mediafilter/model/Page.java
@@ -1,9 +1,16 @@
+/**
+ * The contents of this file are subject to the license and copyright
+ * detailed in the LICENSE and NOTICE files at the root of the source
+ * tree and available online at
+ *
+ * http://www.dspace.org/license/
+ */
 package org.dspace.app.mediafilter.model;
 
 public class Page {
 
-  private int pageNumber;
-  private String text;
+  private final int pageNumber;
+  private final String text;
 
   public Page(int pageNumber, String text) {
     this.pageNumber = pageNumber;
@@ -14,15 +21,8 @@ public class Page {
     return pageNumber;
   }
 
-  public void setPageNumber(final int pageNumber) {
-    this.pageNumber = pageNumber;
-  }
-
   public String getText() {
     return text;
   }
 
-  public void setText(final String text) {
-    this.text = text;
-  }
 }
