+++ dspace-api/src/main/java/org/dspace/app/mediafilter/model/Pages.java
@@ -1,3 +1,10 @@
+/**
+ * The contents of this file are subject to the license and copyright
+ * detailed in the LICENSE and NOTICE files at the root of the source
+ * tree and available online at
+ *
+ * http://www.dspace.org/license/
+ */
 package org.dspace.app.mediafilter.model;
 
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper;
@@ -20,7 +27,4 @@ public class Pages {
     return pageList;
   }
 
-  public void setPageList(final List<Page> page) {
-    this.pageList = page;
-  }
 }
